requirements.txt
# ============================================================
# Core Vanna and FastAPI Components
# ============================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
vanna>=2.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6

# ============================================================
# Memory and Vector Database (Agent Memory)
# ============================================================
chromadb>=0.5.0

# ============================================================
# LLM Providers and API Clients
# ============================================================
openai 
groq 
google-generativeai 

# ============================================================
# Database Connectors 
# ============================================================
pyodbc>=5.0.0      # Required for MSSQL
cx_Oracle>=8.3.0   # Required for Oracle

# ============================================================
# Other Dependencies
# ============================================================
requests
pandas
sqlalchemy